Add rover/base skeleton with Bluetooth support
This patch adds the top-level skeleton code for the rover (src/rover.py)
and the base station (src/base.py). The skeleton code only properly
initializes and shuts down the UBX manager.

On the rover, the skeleton code also creates a Bluetooth transmitter
that can transmit NMEA messages to a remote device, e.g., an Android
device running the Bluetooth GNSS app. The OS must be properly
configured for Bluetooth RFCOMM for this to work and the device must be
paired.